pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simple-gpt-mcp-server"
version = "0.1.0"
description = "MCP Tool Gateway for Memory, CRM, Automation, Inbox and File services"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.23.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
    # Runtime dependencies for FastAPI/uvicorn HTTP app
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.35.0",
    "starlette>=0.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.14.8",
    "mypy>=1.19.0",
    "types-PyYAML",
]

[project.scripts]
simple-gpt-mcp = "mcp_server.main:main"

[tool.setuptools]
packages = ["mcp_server"]
